✨ feat: 링크 생성 뷰와 생성 결과 뷰를 하나로 합치기
