+++ src/presentation/components/NeogaCreateCard/List/index.tsx
@@ -25,6 +25,7 @@ function NeogaCreateCardList(props: NeogaCreateCardListProps) {
           key={card.id}
           {...card}
           onClick={() => {
+            console.log('NeogaCreateCardList', card.isCreated);
             onItemClick(card.id, card.isCreated);
           }}
           idx={idx}
