+++ src/presentation/pages/Neoga/Create/index.tsx
@@ -41,7 +41,7 @@ function NeogaCreate() {
           <NeogaCreateCardList
             cards={allTemplateList}
             onItemClick={(id, isCreated) => {
-              navigate(isCreated ? `/neoga/create/${id}/created` : `/neoga/create/${id}`);
+              navigate(isCreated ? `/neoga/create/${id}/created` : `/neoga/create/${id}/new`);
             }}
           />
         )}
